LA cleanup (#418)
* RawStore cleanup

* MatrixDecomposition

* Elements

* DateContext

* IndexOf

Update CHANGELOG.md

+++ src/main/java/org/ojalgo/matrix/decomposition/LUDecomposition.java
@@ -283,7 +283,7 @@ abstract class LUDecomposition<N extends Comparable<N>> extends InPlaceDecomposi
 
             if (pivoting) {
                 // Find next pivot row
-                int tmpPivotRow = (int) tmpInPlace.indexOfLargestInColumn(ij, ij);
+                int tmpPivotRow = tmpInPlace.indexOfLargestInColumn(ij, ij);
 
                 // Pivot?
                 if (tmpPivotRow != ij) {
